ReadPos/Slice are no longer rc specific, move to olio::fs
